Do not match words for jshint
Think this works better for jshint. Now it will match things like the end ```)``` in ```function(unusedArg)``` and the ```3``` in ```if (5==3)``` and the empty space after the unused variable foo: ```var foo = 5;```

At least this works better when showing the errors in the popup and placing the cursor at the right column.